+++ R/00-AllClasses.R
@@ -1,30 +1,3 @@
-##' Class "compareClusterResult"
-##' This class represents the comparison result of gene clusters by GO
-##' categories at specific level or GO enrichment analysis.
-##'
-##'
-##' @name compareClusterResult-class
-##' @aliases compareClusterResult-class show,compareClusterResult-method
-##'   summary,compareClusterResult-method plot,compareClusterResult-method
-##' @docType class
-##' @slot compareClusterResult cluster comparing result
-##' @slot geneClusters a list of genes
-##' @slot fun one of groupGO, enrichGO and enrichKEGG
-##' @slot .call function call
-##' @exportClass compareClusterResult
-##' @author Guangchuang Yu \url{https://guangchuangyu.github.io}
-##' @exportClass compareClusterResult
-##' @seealso \code{\linkS4class{groupGOResult}}
-##'   \code{\linkS4class{enrichResult}} \code{\link{compareCluster}}
-##' @keywords classes
-setClass("compareClusterResult",
-         representation = representation(
-             compareClusterResult = "data.frame",
-             geneClusters = "list",
-             fun = "character",
-             .call = "call"
-         )
-         )
 
 
 ##' Class "groupGOResult"